Move Relayer signing to helper in Relayer module.
